Refactoring compiling the fw and log context
