+++ fw/arc/log/default.yaml
@@ -1,20 +1,3 @@
-
-# MCORE_BASE = 0x10078000
-# SCRATCH_BASE = 0x80030000
-# LOGGER_ADDRESSES = {
-#     "HEARTBEAT": MCORE_BASE + 0x6c4,
-#     "POWER": MCORE_BASE + 0x18,
-#     "CURRENT": MCORE_BASE + 0x14,
-#     "CURR_AICLK": MCORE_BASE + 0x270,
-#     "TOTAL_TIMER_IRQS": MCORE_BASE + 0x6c8,
-#     "TS_AVG": MCORE_BASE + 0x388,
-#     "VOLTAGE_AT_FMAX": MCORE_BASE + 0x330,
-#     "VOLTAGE": MCORE_BASE + 0x10,
-#     "SCRATCH2": SCRATCH_BASE + 0x68,
-#     "SCRATCH3": SCRATCH_BASE + 0x6C,
-#     "SCRATCH4": SCRATCH_BASE + 0x70,
-#     "SCRATCH5": SCRATCH_BASE + 0x74
-# }
 base_addresses:
   MCORE_BASE: 0x10078000
   SCRATCH_BASE: 0x80030000
@@ -74,16 +57,14 @@ record_configurations:
     output: int
 
 logger_configuration:
-  records: # These will be logged into the circular buffer
-    # - scratch1
-    # - scratch2
-    # - scratch3
-    # - scratch4
-    # - scratch5
-    # - heartbeat
+  default:
     - power
     - current
     - curr_aiclk
     - ts_avg
     - voltage_at_fmax
     - voltage
+  power:
+    - power
+  current:
+    - current
